First commit.
This commit contains 2 NetBeans project. One is the JMapProjLib project from the web, which is a map projection library, the other is the project I created. In order for this to work, both projects need to be situated in the same folder, since there is a library dependancy defined by a relative path.

The code parses the contents of a OSM xml-file. I will send instructions how to get such an xml-file tommorow.

At this moment, junctions are not being parsed yet and therefore not being drawn yet. What you see in the frame are purely road segments.

When you use my xml-file, you will see part of the Brusselse Poort area in Maastricht.
The component has been made such that it automatically resizes with the frame.

The code is not yet documented. I will get to that tommorow, but since I promised to commit the code already last week, I felt I couldn't wait any longer.

===INSTRUCTIONS TO GET STUFF WORKING===
In MainFrame.java I hardcoded the location of an xml-file. I will commit this file as well.
All you have to do is download that file to your pc and substitute its path into the code.